feat: set command

+++ remu_cli/src/tracer.rs
@@ -3,7 +3,7 @@ use remu_types::Tracer;
 pub struct CLITracer;
 
 impl Tracer for CLITracer {
-    fn mem_print(&self, begin: u64, data: &[u8], result: Result<(), Box<dyn std::error::Error>>) {
+    fn mem_print(&self, begin: usize, data: &[u8], result: Result<(), Box<dyn std::error::Error>>) {
         print!("begin: {} ", begin);
         match result {
             Ok(_) => println!("Value: {:?}", data),
