+++ remu_types/src/lib.rs
@@ -1,7 +1,7 @@
 use std::{cell::RefCell, error::Error, rc::Rc};
 
 pub trait Tracer {
-    fn mem_print(&self, begin: u64, data: &[u8], result: Result<(), Box<dyn Error>>);
+    fn mem_print(&self, begin: usize, data: &[u8], result: Result<(), Box<dyn Error>>);
 }
 
 pub type TracerDyn = Rc<RefCell<dyn Tracer>>;
